re: #3135
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@8137 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/cybertracker/org.wcs.smart.cybertracker.survey/src/org/wcs/smart/cybertracker/survey/export/SurveyPackageExporter.java
@@ -235,6 +235,8 @@ public enum SurveyPackageExporter {
 			
 			String fileName = cmObject.getImageFile() == null? cmObject.getDefaultImageFileName() : cmObject.getImageFile().getFileName().toString();			
 			Path toPath = tempDir.resolve(SharedUtils.getFilenameWithoutExtension(fileName) + "." + SharedUtils.getFilenameExtension(fromPath.getFileName().toString())); //$NON-NLS-1$
+
+			if (Files.exists(toPath)) return;
 			Files.copy(fromPath, toPath);
 			if (!toIncludeInZip.contains(toPath)) toIncludeInZip.add(toPath);
 		}
